Execute Do/While loop condition in the proper place
Summary:
Release note: Execute Do/While loop condition in the proper place

Fixes issue: #1677

The fixpoint computation computed the while condition of the first iteration as part of the second, and so on.

Refactored the code to make the fixed point computation independent of where the loop condition is executed. This simplifies things and will come in handy when doing while/do.
Closes https://github.com/facebook/prepack/pull/1741

Differential Revision: D7610522

Pulled By: hermanventer

fbshipit-source-id: 5a86b1ea179b5e03c773b642375b91ad3837d868